chore: extract common linked block code
