+++ packages/rich-text/src/plugins/EmbeddedEntityBlock/LinkedEntityBlock.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 
-import { css } from 'emotion';
 import { useSelected, useReadOnly } from 'slate-react';
 
 import { useContentfulEditor } from '../../ContentfulEditorProvider';
-import { IS_CHROME } from '../../helpers/environment';
 import { findNodePath } from '../../internal/queries';
 import { removeNodes } from '../../internal/transforms';
 import { Element, RenderElementProps } from '../../internal/types';
@@ -12,20 +10,7 @@ import { useSdkContext } from '../../SdkProvider';
 import { useLinkTracking } from '../links-tracking';
 import { FetchingWrappedAssetCard } from '../shared/FetchingWrappedAssetCard';
 import { FetchingWrappedEntryCard } from '../shared/FetchingWrappedEntryCard';
-
-const styles = {
-  root: css({
-    marginBottom: '1.25rem !important',
-    display: 'block',
-  }),
-  container: css({
-    // The next 2 properties ensure Entity card won't be aligned above
-    // a list item marker (i.e. bullet)
-    display: 'inline-block',
-    verticalAlign: 'text-top',
-    width: '100%',
-  }),
-};
+import { LinkedBlockWrapper } from '../shared/LinkedBlockWrapper';
 
 type LinkedEntityBlockProps = {
   element: Element & {
@@ -64,47 +49,39 @@ export function LinkedEntityBlock(props: LinkedEntityBlockProps) {
   }, [editor, element]);
 
   return (
-    <div
-      {...attributes}
-      className={styles.root}
-      data-entity-type={entityType}
-      data-entity-id={entityId}
-      // COMPAT: This makes copy & paste work for Firefox
-      contentEditable={IS_CHROME ? undefined : false}
-      draggable={IS_CHROME ? true : undefined}
+    <LinkedBlockWrapper
+      attributes={attributes}
+      card={
+        <>
+          {entityType === 'Entry' && (
+            <FetchingWrappedEntryCard
+              sdk={sdk}
+              entryId={entityId}
+              locale={sdk.field.locale}
+              isDisabled={isDisabled}
+              isSelected={isSelected}
+              onRemove={handleRemoveClick}
+              onEdit={handleEditClick}
+              onEntityFetchComplete={onEntityFetchComplete}
+            />
+          )}
+          {entityType === 'Asset' && (
+            <FetchingWrappedAssetCard
+              sdk={sdk}
+              assetId={entityId}
+              locale={sdk.field.locale}
+              isDisabled={isDisabled}
+              isSelected={isSelected}
+              onRemove={handleRemoveClick}
+              onEdit={handleEditClick}
+              onEntityFetchComplete={onEntityFetchComplete}
+            />
+          )}
+        </>
+      }
+      element={element}
     >
-      <div
-        // COMPAT: This makes copy & paste work for Chromium/Blink browsers and Safari
-        contentEditable={IS_CHROME ? false : undefined}
-        draggable={IS_CHROME ? true : undefined}
-        className={styles.container}
-      >
-        {entityType === 'Entry' && (
-          <FetchingWrappedEntryCard
-            sdk={sdk}
-            entryId={entityId}
-            locale={sdk.field.locale}
-            isDisabled={isDisabled}
-            isSelected={isSelected}
-            onRemove={handleRemoveClick}
-            onEdit={handleEditClick}
-            onEntityFetchComplete={onEntityFetchComplete}
-          />
-        )}
-        {entityType === 'Asset' && (
-          <FetchingWrappedAssetCard
-            sdk={sdk}
-            assetId={entityId}
-            locale={sdk.field.locale}
-            isDisabled={isDisabled}
-            isSelected={isSelected}
-            onRemove={handleRemoveClick}
-            onEdit={handleEditClick}
-            onEntityFetchComplete={onEntityFetchComplete}
-          />
-        )}
-      </div>
       {children}
-    </div>
+    </LinkedBlockWrapper>
   );
 }
